hilldiv3: Integral Analysis of Diversity Based on Hill Numbers

Tools for analysing the diversity of biological communities (e.g. OTU/ASV/MAG tables) based on Hill numbers. 'hilldiv3' provides a unified framework for neutral, phylogenetic and functional diversity measurement, diversity partitioning, (dis)similarity measurement, diversity profiles, evenness and redundancy. The statistical framework encompasses richness, Shannon and Simpson diversity, Faith's PD, Rao's quadratic entropy and Sorensen/Unifrac-type dissimilarities, all grounded in a single Hill-number framework. Methods are described in Jost (2007) doi:10.1890/06-1736.1 , Chao et al. (2010) doi:10.1098/rstb.2010.0272 , Chiu et al. (2014) doi:10.1890/12-0960.1 and reviewed in Alberdi & Gilbert (2019) doi:10.1111/1755-0998.13014 .

See also

Author

Maintainer: Antton Alberdi antton.alberdi@bio.ku.dk (ORCID)